Change coordinates and direction to f32
We need to perform several numeric calculations. I started to add
different numeric types per context, but that was proving troublesome
because in rust we can't easily make numeric operations with different
types.

This patch changes the coordinate and direction to f32. It also had to
add new logic to check equility because comparing floats has an error
margin.