Refactor world size representation
The integer numbers we were using on `world.size_x` and `world.size_y`
were cumbersome because we had to convert to `f32` many times. To make
this more standard and easy to use, this patch adds a `left_edge` and a
`right_edge`, as coordinates. This also makes the code that checks that
entities are inside the world map simpler.