Remove result responsibility from the world
`get_entity` was returning a result. But if we consider that the world
is just data, then if an entity doesn't exist, that's not an _error_,
that's just state.

The logic that make sure that if we don't have an entity we should have
an error should be moved to the logic layer. Mainly on the movement code
that needs to fetch an entity that should exist.