Add PlayerController
The idea is that actions are _small_ operations that the engine can
process. The _controller_ is responsible for registering the actions
that will be processed for an entity on a given turn.

This patch refactors the code to use a `PlayerController` what will map
a direction to an action. Currently it doesn't do anything new, but the
next step will be to perform an attack instead of a move if the enemy is
in range.

Other controllers will be AI implementation for other entity types
(enemies, buildings, ...).