Refactor input and controller interaction
Moved the player input logic to the world. The point is that the FE
should just register the player input on the world and doesn't need to
know about controllers. Then the `runner` will first process the
controllers, which in turn will generate actions to be processed.

I felt the need for this refactor at this moment to prepare for the next
task: adding enemy controllers. That will be another step on the
`runner`.